change lib, various bug fixes, trails work

+++ src/effect.hh
@@ -20,7 +20,7 @@ class Effect {
 public:
     explicit Effect(Region &region);
     virtual void init() = 0;
-    virtual int16_t loop() = 0;
+    virtual uint64_t loop() = 0;
 
     const PT_THREAD(run(struct pt *proto));
     const void kill();
@@ -30,5 +30,5 @@ protected:
 
 private:
     Waiter w;
-    int16_t delay;
+    uint32_t delay;
 };
