+++ src/effects/static_color.hh
@@ -11,7 +11,7 @@ public:
     explicit StaticColor(Region &region, Color color, uint16_t delay);
 
     void init() override;
-    int16_t loop() override;
+    uint64_t loop() override;
 
 private:
     Color color;
